requirements.txt
python-telegram-bot==22.6
aiohttp>=3.9.0
python-dotenv>=1.0.0
